removed unnecessary casts

+++ db4oj/core/src/com/db4o/io/crypt/XTEA.java
@@ -111,12 +111,12 @@ public class XTEA {
 	 * 
 	 */
 	private void byte2int(byte[] bytes, int offset, int[] res) {
-		res[0] = (int) ((((int) bytes[offset] & 0xff) << 24)
-				| (((int) bytes[offset + 1] & 0xff) << 16)
-				| (((int) bytes[offset + 2] & 0xff) << 8) | ((int) bytes[offset + 3] & 0xff));
-		res[1] = (int) ((((int) bytes[offset + 4] & 0xff) << 24)
-				| (((int) bytes[offset + 5] & 0xff) << 16)
-				| (((int) bytes[offset + 6] & 0xff) << 8) | ((int) bytes[offset + 7] & 0xff));
+		res[0] = (((bytes[offset] & 0xff) << 24)
+				| ((bytes[offset + 1] & 0xff) << 16)
+				| ((bytes[offset + 2] & 0xff) << 8) | (bytes[offset + 3] & 0xff));
+		res[1] = (((bytes[offset + 4] & 0xff) << 24)
+				| ((bytes[offset + 5] & 0xff) << 16)
+				| ((bytes[offset + 6] & 0xff) << 8) | (bytes[offset + 7] & 0xff));
 	}
 
 	/**
